requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==0.11.1
Flask-Cors==3.0.8
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pyowm==2.8.0
requests==2.18.4
six==1.13.0
urllib3==1.22
Werkzeug==0.16.0
wincertstore==0.2
gunicorn==19.9.0
python-dotenv